move ShapedOreEnergyTransferRecipe to common.crafting
